Split out route definition.
